MINIFICPP-1202 - Extract TryMoveCall and void_t, remove redundant checks on running from ConcurrentQueue
